requirements.txt
astropy
astunparse
certifi
chardet
cloudpickle
configparser
csbdeep
dask
dill
h5py
wget
flatbuffers
idna
fpdf
gast
google
httplib2
nodejs
imageio
joblib
jupyterlab
jlab-enhanced-cell-toolbar
# imjoy-jupyter-extension
llvmlite
multiprocess
numba
numpy
oauth2client
py
pyasn1
pydot
requests
rsa
ruamel.yaml
scikit-image
scikit-learn
scipy
six
sortedcontainers
tblib
tensorboard
tensorflow
termcolor
tifffile
toolz
tqdm
uritemplate
urllib3
wrapt
